Decompose SettingsPanel.tsx into sub-components and custom hook
Extract useSettingsState hook for all state/effects/actions,
CrashRecoveryBanner and ConfigSummary components, and engine
mode resolution utilities into shared.ts. SettingsPanel reduced
from 703 to ~160 lines of pure layout.
