+++ src/renderer/components/settings/index.ts
@@ -6,3 +6,5 @@ export { SubtitleSettings } from './SubtitleSettings'
 export { SessionControls } from './SessionControls'
 export { Section } from './Section'
 export { UpdateStatus } from './UpdateStatus'
+export { CrashRecoveryBanner } from './CrashRecoveryBanner'
+export { ConfigSummary } from './ConfigSummary'
